requirements.txt
api==0.0.7
asgiref==3.5.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.4
django-api==0.1.3
django-cors-headers==3.12.0
django-extensions==3.1.5
django-import-export==2.8.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
et-xmlfile==1.1.0
graphviz==0.20
idna==3.3
MarkupPy==1.14
nose==1.3.7
odfpy==1.4.1
openpyxl==3.0.9
Pillow==9.1.1
pydot==1.4.2
pydotplus==2.0.2
PyJWT==1.7.1
pyparsing==3.0.8
pytz==2022.1
PyYAML==6.0
requests==2.27.1
sqlparse==0.4.2
tablib==3.2.1
tzdata==2022.1
urllib3==1.26.9
xlrd==2.0.1
xlwt==1.3.0